MVP glossary tooltips
https://github.com/EFForg/sec/issues/415
Eventually, we'd like the popover to include the first 20 words or so of
the term's definition.  For now, just let the user know that you can
click the thing.

Foundation is supposed to have a dropdown widget that behaves like a
popover, but i couldn't get it to work, and the functionality is simple
enough that i didn't think it was worth sinking more time into figuring
out why.

Improvements:
* Include some or all of the definition in the popover
* Open the popover on click
* Link to the definition inside the popover (currently not available,
because i'm creating the popover as a child of the link, and you can't
have a link inside a link)